+++ packages/saas-ui-react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @saas-ui/react
 
+## 3.0.0-next.26
+
+### Patch Changes
+
+- Fixed util imports
+
+## 3.0.0-next.25
+
+### Patch Changes
+
+- Fixed missing imports
+
+## 3.0.0-next.24
+
+### Patch Changes
+
+- Fix one more import, promised
+
+## 3.0.0-next.23
+
+### Patch Changes
+
+- fc4a915: Fixed missing import extensions
+
 ## 3.0.0-next.22
 
 ### Patch Changes